main frame done

+++ Real-time 2-D Object Recognition/utils/csv_util.cpp
@@ -111,7 +111,7 @@ int getfloat(FILE *fp, float *v) {
 
   The function returns a non-zero value in case of an error.
  */
-int append_image_data_csv( char *filename, char *image_filename, std::vector<float> &image_data, int reset_file ) {
+int append_image_data_csv( char *filename, char *image_filename, std::vector<double> &image_data, int reset_file ) {
   char buffer[256];
   char mode[8];
   FILE *fp;
@@ -158,7 +158,7 @@ int append_image_data_csv( char *filename, char *image_filename, std::vector<flo
 
   The function returns a non-zero value if something goes wrong.
  */
-int read_image_data_csv( char *filename, std::vector<char *> &filenames, std::vector<std::vector<float>> &data, int echo_file ) {
+int read_image_data_csv( char *filename, std::vector<char *> &filenames, std::vector<std::vector<double>> &data, int echo_file ) {
   FILE *fp;
   float fval;
   char img_file[256];
